Converted extraAmounts array from float[] to int[]

+++ src/main/java/com/drakmyth/minecraft/manufactory/recipes/GrinderRecipe.java
@@ -17,7 +17,7 @@ public class GrinderRecipe extends ManufactoryRecipe {
 
     public static final IRecipeType<GrinderRecipe> recipeType = IRecipeType.register("manufactory:grinder");
 
-    public GrinderRecipe(ResourceLocation recipeId, Ingredient ingredient, ItemStack result, float extraChance, float[] extraAmounts, int powerRequired, int processTime) {
+    public GrinderRecipe(ResourceLocation recipeId, Ingredient ingredient, ItemStack result, float extraChance, int[] extraAmounts, int powerRequired, int processTime) {
         super(recipeId, ingredient, result, extraChance, extraAmounts, powerRequired, processTime);
     }
